requirements.txt
faiss-cpu
sentence-transformers
langchain
fastapi
uvicorn
openai>=1.0.0
